[hdf5] move Value writing code to Property
