+++ backend/hdf5/h5x/H5DataSet.cpp
@@ -13,21 +13,6 @@
 #include <cmath>
 
 namespace nix {
-
-
-//special to_data_type for our purporses (Value)
-template<>
-struct to_data_type<char *> {
-    static const bool is_valid = true;
-    static const DataType value = DataType::String;
-};
-
-template<>
-struct to_data_type<const char *> {
-    static const bool is_valid = true;
-    static const DataType value = DataType::String;
-};
-
 namespace hdf5 {
 
 DataSet::DataSet(hid_t hid)
@@ -278,202 +263,6 @@ DataSpace DataSet::getSpace() const {
     DataSpace space = H5Dget_space(hid);
     space.check("DataSet::getSpace(): Could not obtain dataspace");
     return space;
-}
-
-/* Value related functions */
-
-template<typename T>
-struct FileValue  {
-
-    T       value;
-
-    double  uncertainty;
-    char   *reference;
-    char   *filename;
-    char   *encoder;
-    char   *checksum;
-
-    //ctors
-    FileValue() {}
-    explicit FileValue(const T &vref) : value(vref) { }
-
-    inline T val() const { return value; }
-};
-
-template<>
-struct FileValue<bool>  {
-
-    unsigned char value;
-
-    double  uncertainty;
-    char   *reference;
-    char   *filename;
-    char   *encoder;
-    char   *checksum;
-
-    //ctors
-    FileValue() {}
-    explicit FileValue(const bool &vref) :
-            value(static_cast<unsigned char>(vref ? 1 : 0)) {
-    }
-
-    inline bool val() const { return value > 0; }
-};
-
-//
-
-template<typename T>
-h5x::DataType h5_type_for_value(bool for_memory)
-{
-    typedef FileValue<T> file_value_t;
-
-    h5x::DataType ct = h5x::DataType::makeCompound(sizeof(file_value_t));
-    h5x::DataType strType = h5x::DataType::makeStrType();
-
-    h5x::DataType value_type = data_type_to_h5(to_data_type<T>::value, for_memory);
-    h5x::DataType double_type = data_type_to_h5(DataType::Double, for_memory);
-
-    ct.insert("value", HOFFSET(file_value_t, value), value_type);
-    ct.insert("uncertainty", HOFFSET(file_value_t, uncertainty), double_type);
-    ct.insert("reference", HOFFSET(file_value_t, reference), strType);
-    ct.insert("filename", HOFFSET(file_value_t, filename), strType);
-    ct.insert("encoder", HOFFSET(file_value_t, encoder), strType);
-    ct.insert("checksum", HOFFSET(file_value_t, checksum), strType);
-
-    return ct;
-}
-
-#if 0 //set to one to check that all supported DataTypes are handled
-#define CHECK_SUPOORTED_VALUES
-#endif
-#define DATATYPE_SUPPORT_NOT_IMPLEMENTED false
-//
-static h5x::DataType h5_type_for_value_dtype(DataType dtype, bool for_memory)
-{
-    switch(dtype) {
-    case DataType::Bool:   return h5_type_for_value<bool>(for_memory);
-    case DataType::Int32:  return h5_type_for_value<int32_t>(for_memory);
-    case DataType::UInt32: return h5_type_for_value<uint32_t>(for_memory);
-    case DataType::Int64:  return h5_type_for_value<int64_t>(for_memory);
-    case DataType::UInt64: return h5_type_for_value<uint64_t>(for_memory);
-    case DataType::Double: return h5_type_for_value<double>(for_memory);
-    case DataType::String: return h5_type_for_value<char *>(for_memory);
-#ifndef CHECK_SUPOORTED_VALUES
-    default: assert(DATATYPE_SUPPORT_NOT_IMPLEMENTED); return h5x::DataType{};
-#endif
-    }
-}
-
-h5x::DataType DataSet::fileTypeForValue(DataType dtype)
-{
-    return h5_type_for_value_dtype(dtype, false);
-}
-
-h5x::DataType DataSet::memTypeForValue(DataType dtype)
-{
-    return h5_type_for_value_dtype(dtype, true);
-}
-
-template<typename T>
-void do_read_value(const DataSet &h5ds, size_t size, std::vector<Value> &values)
-{
-    h5x::DataType memType = h5_type_for_value<T>(true);
-
-    typedef FileValue<T> file_value_t;
-    std::vector<file_value_t> fileValues;
-
-    fileValues.resize(size);
-    values.resize(size);
-
-    h5ds.read(memType.h5id(), fileValues.data());
-
-    std::transform(fileValues.begin(), fileValues.end(), values.begin(), [](const file_value_t &val) {
-            Value temp(val.val());
-            temp.uncertainty = val.uncertainty;
-            temp.reference = val.reference;
-            temp.filename = val.filename;
-            temp.encoder = val.encoder;
-            temp.checksum = val.checksum;
-            return temp;
-        });
-
-    h5ds.vlenReclaim(memType, fileValues.data());
-}
-
-void DataSet::read(std::vector<Value> &values) const
-{
-    DataType dtype = dataType();
-    NDSize shape = size();
-
-    if (shape.size() < 1 || shape[0] < 1) {
-        return;
-    }
-
-    assert(shape.size() == 1);
-    size_t nvalues = nix::check::fits_in_size_t(shape[0], "Can't resize: data to big for memory");
-
-    switch (dtype) {
-    case DataType::Bool:   do_read_value<bool>(*this, nvalues, values);     break;
-    case DataType::Int32:  do_read_value<int32_t>(*this, nvalues, values);  break;
-    case DataType::UInt32: do_read_value<uint32_t>(*this, nvalues, values); break;
-    case DataType::Int64:  do_read_value<int64_t>(*this, nvalues, values);  break;
-    case DataType::UInt64: do_read_value<uint64_t>(*this, nvalues, values); break;
-    case DataType::String: do_read_value<char *>(*this, nvalues, values);   break;
-    case DataType::Double: do_read_value<double>(*this, nvalues, values);   break;
-#ifndef CHECK_SUPOORTED_VALUES
-    default: assert(DATATYPE_SUPPORT_NOT_IMPLEMENTED);
-#endif
-    }
-
-}
-
-#define NOT_IMPLEMENTED 1
-
-template<typename T>
-void do_write_value(DataSet &h5ds, const std::vector<Value> &values)
-{
-    typedef FileValue<T> file_value_t;
-    std::vector<file_value_t> fileValues;
-
-    fileValues.resize(values.size());
-
-    std::transform(values.begin(), values.end(), fileValues.begin(), [](const Value &val) {
-            file_value_t fileVal(val.get<T>());
-            fileVal.uncertainty = val.uncertainty;
-            fileVal.reference = const_cast<char *>(val.reference.c_str());
-            fileVal.filename = const_cast<char *>(val.filename.c_str());
-            fileVal.encoder = const_cast<char *>(val.encoder.c_str());
-            fileVal.checksum = const_cast<char *>(val.checksum.c_str());
-
-            return fileVal;
-        });
-
-    h5x::DataType memType = h5_type_for_value<T>(true);
-    h5ds.write(memType.h5id(), fileValues.data());
-}
-
-void DataSet::write(const std::vector<Value> &values)
-{
-    setExtent(NDSize{values.size()});
-
-    if (values.size() < 1) {
-        return; //nothing to do
-    }
-
-    switch(values[0].type()) {
-
-    case DataType::Bool:   do_write_value<bool>(*this, values); break;
-    case DataType::Int32:  do_write_value<int32_t>(*this, values); break;
-    case DataType::UInt32: do_write_value<uint32_t>(*this, values); break;
-    case DataType::Int64:  do_write_value<int64_t>(*this, values); break;
-    case DataType::UInt64: do_write_value<uint64_t>(*this, values); break;
-    case DataType::String: do_write_value<const char *>(*this, values); break;
-    case DataType::Double: do_write_value<double>(*this, values); break;
-#ifndef CHECK_SUPOORTED_VALUES
-    default: assert(DATATYPE_SUPPORT_NOT_IMPLEMENTED);
-#endif
-    }
-
 }
